+++ packages/react-native-sdk/src/utils/push/internal/ios.ts
@@ -1,10 +1,8 @@
-import { getLogger, RxUtils } from '@stream-io/video-client';
+import { getLogger } from '@stream-io/video-client';
 import { AppState, NativeModules, Platform } from 'react-native';
 import { getCallKeepLib, getVoipPushNotificationLib } from '../libs';
-import {
-  pushUnsubscriptionCallbacks$,
-  voipPushNotificationCallCId$,
-} from './rxSubjects';
+import { voipPushNotificationCallCId$ } from './rxSubjects';
+import { pushUnsubscriptionCallbacks } from './constants';
 import { canAddPushWSSubscriptionsRef, shouldCallBeEnded } from './utils';
 import { StreamVideoConfig } from '../../StreamVideoRN/types';
 
@@ -46,12 +44,20 @@ export const onVoipNotificationReceived = async (
   }
   const logger = getLogger(['setupIosVoipPushEvents']);
   const client = await pushConfig.createStreamVideoClient();
+
   if (!client) {
     logger(
       'debug',
       'client not found, not processing call.ring voip push notification',
     );
     return;
+  }
+  const shouldRejectCallWhenBusy = client['rejectCallWhenBusy'] ?? false;
+  if (shouldRejectCallWhenBusy) {
+    // inform the iOS native module that we should reject call when busy
+    NativeModules.StreamVideoReactNative.setShouldRejectCallWhenBusy(
+      shouldRejectCallWhenBusy,
+    );
   }
   const callFromPush = await client.onRingingCall(call_cid);
   let uuid = '';
@@ -116,12 +122,9 @@ export const onVoipNotificationReceived = async (
         unsubscribe();
       }
     });
-    const unsubscriptionCallbacks =
-      RxUtils.getCurrentValue(pushUnsubscriptionCallbacks$) ?? [];
-    pushUnsubscriptionCallbacks$.next([
-      ...unsubscriptionCallbacks,
-      unsubscribe,
-    ]);
+
+    pushUnsubscriptionCallbacks.get(call_cid)?.forEach((cb) => cb());
+    pushUnsubscriptionCallbacks.set(call_cid, [unsubscribe]);
   }
   // send the info to this subject, it is listened by callkeep events
   // callkeep events will then accept/reject the call